requirements.txt
Package    Version
---------- ---------
certifi    2022.12.7
cffi       1.15.1
pip        22.3.1
pycparser  2.21
PyNaCl     1.5.0
setuptools 65.6.3
wheel      0.37.1
